refactor(ai): remove @/lib/ai/core barrel — replace with direct submodule imports
Replace all 21 production files + 5 test files that imported from the
@/lib/ai/core barrel with explicit imports from the source submodule
(provider-factory, message-utils, ai-tools, timestamp-utils, etc.).
Delete apps/web/src/lib/ai/core/index.ts.
