+++ apps/web/src/app/api/pages/[pageId]/agent-config/route.ts
@@ -4,7 +4,7 @@ import { canUserEditPage } from '@pagespace/lib/permissions/permissions';
 import { db } from '@pagespace/db/db'
 import { eq } from '@pagespace/db/operators'
 import { pages, drives } from '@pagespace/db/schema/core';
-import { pageSpaceTools } from '@/lib/ai/core';
+import { pageSpaceTools } from '@/lib/ai/core/ai-tools';
 import { validateAgentModelSelection } from '@/lib/ai/core/ai-providers-config';
 import { loggers } from '@pagespace/lib/logging/logger-config'
 import { auditRequest } from '@pagespace/lib/audit/audit-log';
