+++ apps/web/src/app/api/pulse/generate/route.ts
@@ -1,17 +1,9 @@
 import { NextResponse } from 'next/server';
 import { generateText } from 'ai';
 import { authenticateRequestWithOptions, isAuthError } from '@/lib/auth';
-import {
-  createAIProvider,
-  isProviderError,
-  buildTimestampSystemPrompt,
-  getUserTimeOfDay,
-  getStartOfTodayInTimezone,
-  isValidTimezone,
-  normalizeTimezone,
-  formatDateInTimezone,
-  BACKGROUND_LIGHT_MODEL,
-} from '@/lib/ai/core';
+import { createAIProvider, isProviderError } from '@/lib/ai/core/provider-factory';
+import { buildTimestampSystemPrompt, getUserTimeOfDay, getStartOfTodayInTimezone, isValidTimezone, normalizeTimezone, formatDateInTimezone } from '@/lib/ai/core/timestamp-utils';
+import { BACKGROUND_LIGHT_MODEL } from '@/lib/ai/core/ai-providers-config';
 import { db } from '@pagespace/db/db'
 import { eq, and, or, lt, gte, ne, desc, inArray, isNotNull, isNull } from '@pagespace/db/operators'
 import { users } from '@pagespace/db/schema/auth'
