--- apps/web/src/lib/ai/core/index.ts
@@ -1,51 +1,0 @@
-/**
- * @module lib/ai/core
- * @description Core AI utilities shared by both Global AI and Page Agents
- *
- * This module contains the foundational infrastructure for PageSpace's AI systems:
- * - Provider factory for AI model connections
- * - Message utilities for conversation handling
- * - System prompt building
- * - Tool filtering and capabilities
- */
-
-// Provider & Model
-export * from './provider-factory';
-export * from './ai-providers-config';
-export * from './model-capabilities';
-
-// Message & Conversation
-export * from './message-utils';
-export * from './conversation-state';
-
-// System Prompts
-export * from './system-prompt';
-export * from './inline-instructions';
-export * from './mention-processor';
-export * from './timestamp-utils';
-export * from './agent-awareness';
-export * from './page-tree-context';
-export * from './personalization-utils';
-
-// Tools
-export * from './ai-tools';
-export * from './tool-filtering';
-export * from './mcp-tool-converter';
-export * from './integration-tool-resolver';
-
-// Utilities
-export * from './ai-utils';
-export * from './schema-introspection';
-export * from './complete-request-builder';
-
-// Stream Abort (server-side registry)
-export * from './stream-abort-registry';
-
-// Stream Pipe Utilities
-export * from './stream-pipe-utils';
-
-// NOTE: Client-side stream-abort functions are NOT exported here to prevent
-// accidental import in server contexts. Import from '@/lib/ai/core/client' instead.
-
-// Types
-export * from './types';
