+++ apps/web/src/lib/channels/agent-mention-responder.ts
@@ -12,7 +12,7 @@ import {
 } from '@/lib/websocket/socket-utils';
 import { processMentionsInMessage } from '@/lib/ai/core/mention-processor';
 import { buildThreadPreview } from '@pagespace/lib/services/preview';
-import type { ToolExecutionContext } from '@/lib/ai/core';
+import type { ToolExecutionContext } from '@/lib/ai/core/types';
 
 const channelMentionLogger = loggers.ai.child({ module: 'channel-agent-mentions' });
